:recycle: [Refactor] openPostDetail 함수 분리 #main
